==> Instacart_Basket_(IC)_Analysis/03 Scripts/2. Importing Libraries and Python Data Types.ipynb <==
# 2. Importing Libraries and Python Data Types
## Table of contents
***

**`1. Import libraries`**<br/>

**`2. Basic calculation`**<br/>

**`3. Division with floating-point variables`** <br/>

**`4. Construct a short word made of separate strings`**<br/>

**`5. Construct 2 short sentences made of separate strings`**<br/>

**`6. Bonus task`**<br/>
***

### 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Basic calculation

#### _3 different ways of reaching a result of 100_

2*50

80+20

10000/100

### 3. Division with floating-point variables

105.78/2.43

### 4. Construct a short word made of separate strings

a = 'Hello,'

b = ' world.'

a+b

### 5. Construct 2 short sentences made of separate strings

c = 'currently'

d= 'learning'

e= 'Python'

a+b+' '+c+' '+d+' '+e

### 6. Bonus Task

'cool' *100
