==> Instacart_Basket_(IC)_Analysis/03 Scripts/3. Data Import and Descriptive Analysis.ipynb <==
# 3. Instacart Basket Data Import and Descriptive Analysis
## Table of contents
***

**`1. Import libraries`**<br/>

**`2. Create reusable path`**<br/>

**`3. Import data`**<br/>

**`4. Create lists, omitting the “eval_set” column`**<br/>

**`5. Obtain the dimension of the dataframe`**<br/>

**`6. Obtain descriptive information`** <br/>

**`7. Obtain dataframe information`**<br/>
***

### 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create Reusable Path (Turing folder path into a string)

# Create reusable path
path = (r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)')

# Check output
path

### 3. Import data using reusable path

df_orders = pd.read_csv(os.path.join(path,'02 Data','Original Data','orders.csv'), index_col=False)

### 4. Create lists, omitting the “eval_set” column

# Check column names
df.columns

# Create list with given columns
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow',
       'order_hour_of_day', 'days_since_prior_order']

# Import data
df_orders = pd.read_csv(os.path.join(path, '02 Data','Original Data','orders.csv'),usecols=vars_list)

df.head()

# Import dataframe
df_prods = pd.read_csv(os.path.join(path, '02 Data','Original Data','products.csv'), index_col = False)

# View data from the beginning of the dataframe
df_prods.head(20)

# View data from the end of the dataframe
df_prods.tail(35)

# Check columns
df_prods.columns

### 5. Obtain the dimension of the dataframe

# Check the shape of the dataframe
df_prods.shape

### 6. Obtain descriptive information

#### _What’s the maximum value of the “aisle_id” column?_

# Check descriptive information of columns
df_prods.describe()

Answer: 134.000000

### 7. Obtain dataframe information

#### _What’s the data type of the “department_id” column?_

# Obtain datafame information
df_prods.info()

Answer: int64
