==> Instacart_Basket_(IC)_Analysis/03 Scripts/4. Data Wrangling and Subsetting.ipynb <==
# 4. Data Wrangling and Subsetting
## Table of contents
***

**`1. Import libraries`**<br/>

**`2. Create path and import data`**<br/>

**`3. Wrangling data`** <br/>
- _3.1 Change data types_<br/>
- _3.2 Rename columns_<br/>
- _3.3 Drop columns_<br/>

 
**`4. Wrangling department dataframe`** <br/>
- _4.1 Transposing department dataframe_ <br/>
- _4.2 Change header of the department dataframe_ <br/>
- _4.3 Create data dictionary_ <br/>

**`5. Locate the busiest hour of day`**<br/>

**`6. Understand breakfast item sales`**<br/>
- _6.1 Create a subset for breakfast item_<br/>
- _6.2 Create a subset for dinner party item_<br/>
- _6.3 Understand the specific user behavior_<br/>

**`7. Export data`**
***

### 1. Import Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create and import data

# Create path
path = r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)'

# Orders data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Products data
df_prods = pd.read_csv(os.path.join(path,'02 Data', 'Original Data', 'products.csv'), index_col = False)

# Departments data
df_dep = pd.read_csv(os.path.join(path,'02 Data', 'Original Data', 'departments.csv'), index_col = False)

### 3. Wrangling data
#### _3.1 Change data types_

# Changing data type for 'user_id' variable from numerical to string
df_ords['user_id'] = df_ords['user_id'].astype('str')

# Changing data type for 'order_id variable from numerical to string'
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Check results
df_ords['order_id'].dtype
df_ords['user_id'].dtype

**Observation**

object type in pandas = string in Python

#### _3.2 Rename columns_

df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)
df_ords.rename(columns= {'order_dow' : 'order_day_of_the_week'}, inplace = True)

df_ords.head()

#### _3.3 Drop columns_

df_ords = df_ords.drop(columns = ['eval_set'])

# Check results
df_ords.head()

**Observation**

After careful consideration, the unnecessary column of 'eval_est' has been dropped succesfully from the dataframe.

## 4. Wrangling department dataframe 

#### _4.1 Transposing the department df_

# Transposing the department data
df_dep_t = df_dep.T

#### _4.2 Change header of the department df_

# Step1: adding an index
df_dep_t.reset_index()

# Step 2: create a new header
new_header = df_dep_t.iloc[0]

# Step 3: create new dataframe to discard the old header
df_dep_t_new = df_dep_t[1:]

# Step 4: set the header rows as the df header
df_dep_t_new.columns = new_header

# Check the final result of header changing
df_dep_t_new

#### _4.3 Create data dictionary_

# Set the new df as data dictionary
data_dict = df_dep_t_new.to_dict('index')

# Check result of the data dictionary
data_dict

**Observation**

The data dictionary is created using the index number to represent a single entry which helps Python locate the information in the dictionary efficiently.

### 5. Locate the busiest hour of day

# Find the busiest hour for placing orders: 10 o'clock
df_ords['order_hour_of_day'].value_counts (dropna= False)

**Observation**

The busiest hour of order being placed is 10 o'clock and the least busy hour is 3 o'clock in the morning. It starts to get busy from 8am and peak at 10am. From 10am till 17pm, the orders slightly decrease but still remain considerably high, however, orders drastically decrease after 18:00pm.

# Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe
data_dict.get('4')

### 6. Understand breakfast item sales

#### _6.1 Create a subset for breakfast item_

# Create a subset for breakfast item, identify the department id first
data_dict

**Observation**

The index represents breakfast department is '14'.

# Create the subset for breakfast item
df_prods_breakfast = df_prods.loc[df_prods['department_id'].isin([14])]

# Check result
df_prods_breakfast.head()

#### _6.2 Create a subset for dinner party item_

# Create 'dinner party' subset
df_prods_dinner = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

# Check result
df_prods_dinner.head()

**Observation**

The dataframe created above for dinner party could provide the sales team a good glance to start their analysis.

# Find out how many rows are in the dataframe last created
df_prods_dinner.shape

**Observation**

The shape of dataframe 'df_dinner_party' shows 7650 rows and 5 columns, meaning 7650 products are included in the list.

#### _6.3 Understand the specific user behavior_

# Extracting all information about user_1
df_user_1 = df_ords.loc[df_ords['user_id']=='1']
df_user_1

# Provide basic stats of the behavior of user_1
df_user_1.describe()

#### Observation:
This user only places orders from Sunday to Wednesday, never from Thrusday to Saturday. 
More than half of the time, he places orders at or before 8am and the orders are placed in average every 19 days. 
He never spend more than 30 days without placing orders, and once he placed 2 orders on the same day.

### 7. Export data

# Export orders dataframe
df_ords.to_csv(os.path.join(path,'02 Data','Prepared Data','orders_wrangled.csv'))

# Export department dataframe
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
