==> Instacart_Basket_(IC)_Analysis/03 Scripts/5. Data Consistency Check.ipynb <==
# 5. Data Consistency Check
## Table of contents

***

**`1. Import libraries`**<br/>

**`2. Create reusable path`**<br/>

**`3. Import data`**<br/>

**`4. Example of identifying mixed-type columns`**<br/>

**`5. Product dataframe consistency check`**<br/>
- _5.1 Locate missing values (NULL)_<br/>
- _5.2 Isolate missing values_<br/>
- _5.3 Address missing values_ <br/>
- _5.4 Isolate the duplicates_<br/>
- _5.5 Remove the duplicates_<br/>

**`6. Order dataframe data consistency check`**<br/>
- _6.1 Investigate the accuracy of all columns_<br/>
- _6.2 Identifying mixed type column_ <br/>
- _6.3 Check for missing values_ <br/>
- _6.4 Address the missing value_<br/>
- _6.5 Check for the duplicate values_<br/>
- _6.6 Addressing dulplicates_<br/>

**`7. Export final dataframe for both 'product' & 'order' dataframes`**
***

### 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create reusable path

# Create path
path = r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)'

### 3. Import Data

# Import 'product.csv' dataset from original data
df_prods = pd.read_csv(os.path.join(path, '02 data' , 'original data', 'products.csv'))

# Import 'orders_wrangled.csv' dataset from prepared data
df_ords = pd.read_csv(os.path.join(path, '02 data' , 'prepared data', 'orders_wrangled.csv'), index_col = 0)

### 4. Example of identifying mixed-type columns

# Create a test dataframe
df_test = pd.DataFrame()

# Create a mixed-type column
df_test ['mix'] = ['a', 'b', 1 , True]

# Check output
df_test.head()

# Identifying mixed-type column within the dataframe 'df_test'
for col in df_test.columns.tolist():
    weird = (df_test[[col]] . applymap(type) != df_test [[col]] . iloc[0] . apply(type)) . any(axis = 1)
    if len (df_test [weird]) > 0 :
             print (col)

### 5. Product dataframe consistency check

#### _5.1 Locate missing value (NULL)_

# Step 1: look for missing value
df_prods.isnull().sum()

#### _5.2 Isolate missing values_

# Step 2: create a subset of the dataframe containing only the NaN values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Check the number of rows 
df_prods.shape

#### _5.3 Address missing value_

<p>method 1: create new variable to flag the missing value
<p>method 2: impute the value with mean or median
<p>method 3: remove the row contains missing values (demonstrated below)

# Step 3: create a new dataframe that filters the missing values ou in 'product_name' column
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

#### _5.4 Isolate the duplicates_

# Step 1: create a subset that only contains the dulplicates
df_dups=df_prods_clean[df_prods_clean.duplicated()]

# Check the number of rows that are duplicated
df_dups.shape

# Before dropping dulplicates, check the number of row of the current dataframe
df_prods_clean.shape

# Check output
df_dups

#### _5.5 Remove the duplicates_

# Drop the duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# Check the row of the dataframe after the removal of the duplicates
df_prods_clean_no_dups.shape

### 6. Order dataframe data consistency check

#### _6.1 Investigate the accuracy of all columns_

# Obtain descriptive information
df_ords.describe()

**interpretation:**

- **order_number**<br>
Maximum order number indicates that the order number had reached 100. <br>
- **order_day_of _week**<br>
Minimum value '0' and maximum value '6' show that there're seven numbers indicating the 7 days of a week.<br>
- **order_hour_of_day**<br>
Minimu value '0' and maximum value '23 total the 24 hours of a day.<br>
- **days_since_last_order**<br>
Minimum value '0' indicates that some orders were placed on the same day; maximum value of '30' shows the 30 days of a month.<br>

**Note:** The counts of column 'days_since_last_order is less than the rest of the columns, indicating there're missing values in the column.<br>
<br>
**Data Type:**<br>
The order_id and user_id should be string data type rather than numerical.<br>

#### _6.2 Identifying mixed type column_

for col in df_ords.columns.tolist():
    weird = (df_ords[[col]] . applymap(type) != df_ords [[col]] . iloc[0] . apply(type)) . any(axis = 1)
    if len (df_ords [weird]) > 0 :
             print (col)

**Obervsation:**<br>
There's no column with mix-type data present in df_ords

# Check the data type of all the columns
df_ords.dtypes

#### _6.3 Check for missing values_

df_ords.isnull().sum()

**Obervation:**<br>
There're '206209' rows with missing values in column of days_since_last_order, the rest of the columns do not have missing values which justify the investigation from column accuracy check.

>**Appropriate method evaluation**<br>
>The column of 'days_since_last_order' would not have any values if the entry is from the first orders, base on this reason, the missing values are considered relevent.<br>
>After careful evaluation, to be able to continue the current analysis, the missing values should be separated from the dataframe by creating a subset that contains only the nulls.

#### _6.4 Address the missing value_

# Step 1: separate the NaN value in the column
df_ords_nan = df_ords[df_ords['days_since_last_order'].isnull() == True]

df_ords_nan

**Observation**<br>
A verification of whether all the entries with NaN value in column 'days_since_last_order' are due to '1' order is needed.

# Step 2: filter the dataframe where 'order_number' equals 1
df_first_orders = df_ords_nan[df_ords_nan['order_number'] == '1']

# Step3: check if all of these have missing values in 'days_since_last_order'
df_first_orders_missing = df_first_orders['days_since_last_order'].isna().all()

# Step 4: validate
df_first_orders_missing

# Print the complete result
print("All first orders have missing value in 'days_since_last_order':", df_first_orders_missing)

# Step 5: Address the missing values by creating a column that flags order numbers as either True/False depending on if the 'order_number' = 1
df_ords['first_order'] = df_ords['days_since_last_order'].isnull()

# Step 6: Check the new inserted column
df_ords

# Step 7: rename df_ords to df_ords_clean
df_ords_clean = df_ords

# Step 8: check data-type of new column
df_ords_clean['first_order'].dtype

# Step 9: double check if the dataframe contains the right number of rows of NaN value
df_ords_clean.isnull().sum()

**Observation**<br>
The NaN values are addressed appropriately, clearly flagged in 'first_order' column which helps the client to understand why the NaN values existed in 'days_since_last_order' column<br>
<br>
The removal of NaN value in 'days_since_last_order' could result in inaccurate analysis as the information is vital for understanding customer behavior.

#### _6.5 Check for the duplicate values_

# Step 1: create a subset of duplicates from df_ords
df_ords_clean_dup = df_ords_clean[df_ords_clean.duplicated()]

# Step 2: check the duplicates in the subset created
df_ords_clean_dup

# Step 3: rename 'df_ords_clean_dup' to 'df_ords_clean_no_dup'
df_ords_clean_no_dups = df_ords_clean

#### _6.6 Addressing dulplicates_

**Observation**<br>
The dataframe created to check duplicates has returned empty, meaning there are no duplicates in the data set.

### 7. Export final dataframe for both 'product' & 'order' dataframes

df_ords_clean_no_dups.to_csv(os.path.join(path, '02 data','prepared data','orders_cleaned.csv'))
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 data', 'prepared data', 'products_cleaned.csv'))
