==> Instacart_Basket_(IC)_Analysis/03 Scripts/6. Data Combining and Exporting.ipynb <==
# 6. Data Combining and Exporting
## Table of contents

***

**`1. Import libraries`**<br/>

**`2. Create reusable path`** <br/>

**`3. Import dataframes`**<br/>

**`4. Understand the basic information of the dataframes`**<br/>

**`5. Merge dataframes`**<br/>
-_5.2 Merge testing_<br/>
-_5.3 Perform actual merge_<br/>

**`6. Export merge data`**<br/>
***

### 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create reusable path

# Create path
path = r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)'

### 3. Import dataframes

# Import dataframe
orders_products_combined = pd.read_pickle(os.path.join(path, '02 data' , 'prepared data' , 'orders_products_combined.pkl'))
df_prods = pd.read_csv(os.path.join(path, '02 data' , 'prepared data' , 'products_cleaned.csv'), index_col = 0)

# Check the shape of the imported dataframe
orders_products_combined.shape

### 4. Understand the basic information of the dataframes

# Check dataframe structure
orders_products_combined.head()

# Check the shape
df_prods.shape

# Check dataframe structure
df_prods.head()

### 5. Merge dataframes

**Evaluation**<br/>
There's a key column 'product_id' observed among these two dataframes, indicating a merge is possible, however, as for which type of the merge should be applied to depends on the information we expect to see in the final dataframe. Since the information wouldn't be much useful if it doesn't exist in both dataframes, therefore, the 'inner' merge will be a suitable opion.

#### _5.2 Merge testing_

# Test merge without overwriting with merge flag 'match'
pd.merge(orders_products_combined , df_prods , on = ['product_id'] , how = 'inner', indicator = 'match' )

**Evaluation** <br/>
As the previous exercise of merged dataframe has created 'merge' column as the flag which would be no use for the new merge of 'orders_products_combined' and 'df_prods', for better presetation, the 'merge' column from the previous merge should be dropped after the new merge has been carried out.

#### _5.3 Perform actual merge_

# Actual merge of 'orders_products_combined' and '_df_pords'
ords_prods_merge = orders_products_combined.merge(df_prods , on = ['product_id'] , indicator = 'match' )

# Drop the flag column '_merge' from the previous merge
ords_prods_merge.drop('_merge', axis = 1 )

### 6. Export merged data

# Export data as pickle
ords_prods_merge.to_pickle(os.path.join(path,'02 data' , 'prepared data' , 'ords_prods_merge.pkl'))
