==> Instacart_Basket_(IC)_Analysis/03 Scripts/7. Deriving New Variables.ipynb <==
# 7. Deriving New Variables
## Table of contents
***

**`1. Import libraries`**<br/>

**`2. Create reusable path`**<br/>

**`3. Import dataframes`**<br/>

**`4. Create a subset`**<br/>

**`5. Identify product price range by user-defined function`**<br/>
- _5.1 Define function on subset_<br/>
- _5.2 Apply function on subset_<br/>
- _5.3 Derive prodcut price range flag by label indexing_<br/>
- _5.4 Apply on entire dataframe_<br/>

**`6. Deriving new variables using If-Statements with For-Loops`**<br/>
- _6.1 Identify busiest day/ least busy day_<br/>

**`7. Export dataframe`**<br/>
***
                      


### 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create reusable path

# Create path
path = r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)'

### 3. Import dataframes

# Import dataframe 'ords_prods_merge'
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 data' , 'prepared data' , 'ords_prods_merge.pkl'))

### 4. Create a subset to operate on (avoid memory issue)

df= ords_prods_merge[:1000000]

df.shape

### 5. Identify product price range by user-defined function

#### _5.1 Define function_

# Define function
def price_label(row):
    
    if row ['prices'] <= 5:
        return 'Low-range product'
    elif (row ['prices'] > 5) and (row ['prices'] <= 15) :
        return 'Med-range product'
    elif row [ 'prices'] > 15:
        return 'High range'
    else : return 'Not enough data'

#### _5.2 Apply function_

# Apply function in the newly created column
df ['price_range'] = df.apply(price_label, axis = 1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

#### _5.3 Derive prodcut price range flag by label indexing_

# Prices larger than 15
df.loc [df ['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Prices less than or equal to 15 and larger than 5 
df.loc [ (df ['prices'] <= 15) & (df ['prices'] > 5) , 'price_range_loc'] = 'Mid-range product'

# Prices less than or equal to 5
df.loc [ df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Check output
df ['price_range_loc'].value_counts (dropna = False)

#### _5.4 Apply on entire dataframe_

# Prices larger than 15
ords_prods_merge.loc [ords_prods_merge ['prices'] > 15 , 'price_range_loc'] = 'High-range product'

# Prices less than or equal to 15 and larger than 5 
ords_prods_merge.loc [ (ords_prods_merge ['prices'] <= 15) & (ords_prods_merge ['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# Prices less than or equal to 5
ords_prods_merge.loc [ ords_prods_merge ['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Check output
ords_prods_merge ['price_range_loc'].value_counts (dropna = False)

### 6. Deriving new variables using if-Statements with For-Loops

#### _6.1 Identify busiest day/ least busy day_

# Get the frequence to know which day most orders take place
ords_prods_merge ['order_day_of_the_week'].value_counts(dropna = False)

**Observation** <br/>
Number of orders has been arranged in descending order, the day '0' with the most orders represent 'Saturday', while day '4' represents Wednesday

# Apply for-loop integrated with if-statements
result = []

for value in ords_prods_merge ['order_day_of_the_week'] :
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else :
        result.append('Regularly busy')

result [:5]

**Observation** <br/>
The list 'result' does not help much, it should be combined with the ords_prods_merge dataframe

# Combine result list with the dataframe
ords_prods_merge ['busiest_day'] = result

# Count the frequency of the new created column and cross-check it with frequency check done before the loop
ords_prods_merge ['busiest_day'].value_counts(dropna = False)

### 7. Export dataframe

# Export data as pickle
ords_prods_merge.to_pickle(os.path.join(path, '02 data' , 'prepared data' , 'ords_prods_derived.pkl'))
