==> Instacart_Basket_(IC)_Analysis/03 Scripts/9. Customer Data Preparation.ipynb <==
# 9. Customer Data Preparation
#### _Clean customer data and combine it with the rest of prepared Instacart data_
## Table of contents
***

**`1. Import libraries`**<br/>

**`2. Create path and import data`**<br/>

**`3. Data wrangling`**<br/>
- _3.1 Column format unification_<br/>
- _3.2 Conver column data types_<br/>
- _3.3 Drop unnecessary columns_<br/>


**`4. Data consistency check`**<br/>
- _4.1 Identify mixed-type columns_<br/>
- _4.2 Locate missing values (NULL)_<br/>
- _4.3 Locate duplicates_<br/>
    
**`5. Data preparation prior to the merge`**<br/>

**`6. Inner merge data`**<br/>

**`7. Export data`**

***

### 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create path and import data

# Create path
path = r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)'

# Import customers data
df_cus = pd.read_csv(os.path.join(path, '02 data' , 'original data' , 'customers.csv'), index_col = False)

# Check data info
df_cus.info()

**Obervation:**<br/>
*`Data wrangling`* <br/>
     - Column formats are not unified.<br/>
     - All variables provide essential information for the analysis, there's no need to drop any columns.<br/>
     - Column 'user_id' and 'age' are considered as categorical variables rather than numerical variables, they should be converted into 'string' values.<br/>
<br/>
*`Data consistency check`*<br/>
    - In column 'First Name', the non-null value count does not match with the rest of the columns. It is suspected that there're null values present in the column that should be investigated.<br/>

# Check data shape
df_cus.head()

### 3. Data wrangling

#### _3.1 Column format unification_

# Rename columns
df_cus = df_cus.rename(columns = {'First Name':'first_name' , 'Surnam':'surname' , 'Gender':'gender' , 'STATE':'state' , 'Age':'age' , 'dependants':'n_dependants', 'fam_status':'marital_status'})

# Check output
df_cus.head()

**Observation:** <br/>
The format of all columns are uniformized.

#### _3.2 Convert column data types_

# Convert data type for column user_id from numerical to string
df_cus['user_id'] = df_cus['user_id'].astype('str')

# Convert data type for column user_id from numerical to string
df_cus['age'] = df_cus['age'].astype('str')

# Check results
df_cus.info()

**Obersvation:**
Data types of 'user_id' and 'age' are converted into 'Object' from 'inte64'. (string in Python)

#### _3.3 Drop unnecessary columns_<br/>
- All variables provide essential information for the analysis, there's no need to drop any columns.

# Downsize data
df_cus = df_cus.astype({col: 'int32' for col in df_cus.select_dtypes('int64').columns})

# Check results
df_cus.info()

### 4. Data consistency check using for-loops

#### *4.1 Identify mixed-type columns*

for col in df_cus.columns.tolist():
    weird = (df_cus[[col]] . applymap(type) != df_cus [[col]] . iloc[0] . apply(type)) . any(axis = 1)
    if len (df_cus [weird]) > 0 :
             print (col)

**Observation:**<br/>
The column 'first_name' contains mixed-type values.

# Convert 'first_name' column data type to 'string'
df_cus['first_name'] = df_cus['first_name'].astype('str')

# Check results
for col in df_cus.columns.tolist():
    weird = (df_cus[[col]] . applymap(type) != df_cus [[col]] . iloc[0] . apply(type)) . any(axis = 1)
    if len (df_cus [weird]) > 0 :
             print (col)

**Observation:**
Mixed-type data in column 'first_name' is fixed.

#### *4.2 Locate missing values*

# Check if there's anynull values in the dataframe
df_cus.isnull().sum()

**Appropriate method evaluation**<br/>
After mixed-type values are fixed in colmun 'first_name', there's no more missing values in the dataframe.

# Cross check the results
df_cus.info()

#### *4.3 Locate duplicates*

# Create a subset that only contains duplicates
df_dups = df_cus[df_cus.duplicated()]

# Check the subset
df_dups

**Observation:**<br/>
There's no duplicates exist in the dataframe.

# Check value frequency in gender column
df_cus['gender'].value_counts(dropna = False)

**Observation:**<br/>
This steps is to make sure whether there's more different types of gender in the column, such as, nonbinary, bisexual, transgender,etc.

### 5. Data preparation prior to the merge

# Import the other prepared Instacart data
ords_prods = pd.read_pickle (os.path.join(path, '02 data' , 'prepared data' , 'ords_prods_aggregated.pkl'))

ords_prods.info()

# Downsize dataframe by converting data type 
ords_prods = ords_prods.astype({col: 'int32' for col in ords_prods.select_dtypes('int64').columns})

# Convert data type 
ords_prods = ords_prods.astype({'order_number':'int16', 'order_day_of_the_week':'int16', 'order_hour_of_day':'int16', 'days_since_last_order':'float16', 'department_id':'int16', 'prices':'float16','aisle_id':'int16', 'max_order':'int16','avg_order_price':'float32','median_of_regularity':'float16'})

# Check output
ords_prods.info()

**Observation:**<br/>
The memory usage of ords_prods dataframe has been successfully downsized from 5.9GB to 3.5GB.

# Check the shape of newly imported dataframe
ords_prods.shape

# Check the shape of df_cus dataframe
df_cus.shape

# Identify key column in both dataframes
ords_prods.head(5)

df_cus.head(5)

**Observation:**
Column 'user_id' is identified as the key column for the merge.

# Check the data type of the key column in ords_prods dataframe
ords_prods['user_id'].dtype

# Check the data type of the key column in df_cus dataframe
df_cus['user_id'].dtype

# Convert the data type of 'user_id' column in ords_prods dataframe
ords_prods['user_id'] = ords_prods['user_id'].astype ('str')

# Check result
ords_prods['user_id'].dtype

**Observation:**<br/>
The data type of the key columns in both dataframes are unified.

### 6. Inner merge dataframes

# Drop the existing '_merge' column from previous exercises
ords_prods = ords_prods.drop(columns=['_merge'])

ords_prods_cus = ords_prods.merge(df_cus, on = 'user_id' , indicator = True)

# Check shape of newly merged dataframe
ords_prods_cus.shape

pd.options.display.max_columns = None

# Check result
ords_prods_cus.head()

# Drop match column from previous merge
ords_prods_cus = ords_prods_cus.drop('match', axis=1)

# Check inner join value counts
ords_prods_cus['_merge'].value_counts()

**Observation:**<br/>
Two dataframes are successfully merged.

### 7. Export dataframe

# Export data as pickle
ords_prods_cus.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_customers_merged.pkl'))
