==> Instacart_Basket_(IC)_Analysis/03 Scripts/10. Part_1_Data Privacy & Low-activity User Exclusion.ipynb <==
# 10. Part 1_Data Privacy & Low-activity User Exclusion
## Table of Contents
***

**`1. Import libraries`**

**`2. Create path and import data`**

**`3. Create an user exclusion for low-activity customers`**

**`4. Export user exclusion data`**

***

### 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.colors as mcolors
import seaborn as sns
import scipy
import os

### 2. Create path and import data

# Create path
path= r'/Users/bonnieandbanana/Desktop/A New Career/CareerFoundry/Data Immersion/A4_Python/ Instacart Basket Analysis Project (IC)'

# Import data
df = pd.read_pickle(os.path.join(path,'02 data' , 'prepared data', 'new_orders_products_customers_merged.pkl'))

pd.options.display.max_columns = None

df['age'] = pd.to_numeric(df['age'])

df.head(5)

**Identifying personally identifiable information (PII):** <br/>
Colmuns of 'first_name', 'surname' are identified as PII.

###  3. Create an exclusion flag: low-activity & high-activity customers

# Create a flag for customers with less than 5 orders
df.loc[ df ['max_order'] <5 , 'activity_flag'] = 'low-activity customer'
df.loc[ df['max_order'] >= 5, 'activity_flag'] = 'high-activity customer'

# Check value counts
df['activity_flag'].value_counts(dropna = False)

# Exclude the low-activity customers from the dataframe
df = df[df['activity_flag'] == 'high-activity customer']

# Check shape
df.shape

**Confirmed that the number of rows match with the rows of high-activity customer**

### 4. Export user exclusion data

# Export the exclusion as pickle
df.to_pickle(os.path.join(path, '02 data' , 'prepared data' , 'low-activity_cus_excluded.pkl'))
